Extract JsonFormTree component into own package
